refactor citation/citationi_item
